requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=1.0.0

# Vector stores
chromadb>=0.4.22

# Document loaders
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
chardet>=5.0.0

# Embeddings
sentence-transformers>=2.3.0
langchain-huggingface>=0.0.1

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
fastapi>=0.115.0
uvicorn>=0.30.0

# Optional: For better text processing
tiktoken>=0.5.0
